requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.57.4
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
